# cctv_server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para sistema de gestión de tickets CCTV multi-tenant",
        "title": "SyMTickets CCTV API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@symtickets.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/intelligence/analyses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of AI analyses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intelligence"
                ],
                "summary": "List AI analyses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by analysis type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handlers.AnalysisResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/intelligence/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of AI model configurations for tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intelligence"
                ],
                "summary": "List AI model configurations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handlers.ModelConfigResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/intelligence/templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of available prompt templates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intelligence"
                ],
                "summary": "List prompt templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handlers.PromptTemplateResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/intelligence/usage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get usage and cost statistics for AI",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intelligence"
                ],
                "summary": "Get AI usage statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UsageStatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.AnalysisResponse": {
            "type": "object",
            "properties": {
                "analysis_type": {
                    "type": "string"
                },
                "confidence_score": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_type": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "model_used": {
                    "type": "string"
                },
                "result": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "internal_handlers.ModelConfigResponse": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "type": "string"
                },
                "default_max_tokens": {
                    "type": "integer"
                },
                "default_temperature": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_default": {
                    "type": "boolean"
                },
                "max_requests_per_day": {
                    "type": "integer"
                },
                "max_tokens_per_request": {
                    "type": "integer"
                },
                "model_name": {
                    "type": "string"
                },
                "monthly_budget_usd": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.PromptTemplateResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "response_format": {
                    "type": "string"
                },
                "system_prompt": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "user_prompt_template": {
                    "type": "string"
                },
                "variables": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handlers.UsageStatsResponse": {
            "type": "object",
            "properties": {
                "failed_calls": {
                    "type": "integer"
                },
                "successful_calls": {
                    "type": "integer"
                },
                "total_api_calls": {
                    "type": "integer"
                },
                "total_cost_usd": {
                    "type": "number"
                },
                "total_tokens": {
                    "type": "integer"
                },
                "usage_by_provider": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}